+++ after_init.rb
@@ -144,10 +144,6 @@ Rails.application.configure do
   Redmine::Activity.register :dmsf_file_revision_accesses, default: false
   Redmine::Activity.register :dmsf_file_revisions
 
-  # Uncomment to remove the original Documents from searching and project's modules (replaced with DMSF)
-  # Redmine::Search.available_search_types.delete('documents')
-  # Redmine::AccessControl.available_project_modules.delete(:documents)
-
   if Redmine::Plugin.installed?('easy_extensions')
     require "#{File.dirname(__FILE__)}/lib/redmine_dmsf/webdav/custom_middleware"
     config.middleware.insert_before ActionDispatch::Cookies, RedmineDmsf::Webdav::CustomMiddleware