+++ lib/redmine_dmsf.rb
@@ -39,6 +39,8 @@ require "#{File.dirname(__FILE__)}/redmine_dmsf/patches/issue_patch"
 require "#{File.dirname(__FILE__)}/redmine_dmsf/patches/role_patch"
 require "#{File.dirname(__FILE__)}/redmine_dmsf/patches/queries_controller_patch"
 require "#{File.dirname(__FILE__)}/redmine_dmsf/patches/pdf_patch"
+require "#{File.dirname(__FILE__)}/redmine_dmsf/patches/access_control_patch"
+require "#{File.dirname(__FILE__)}/redmine_dmsf/patches/search_patch"
 
 # A workaround for obsolete 'alias_method' usage in RedmineUp's plugins
 if RedmineDmsf::Plugin.an_obsolete_plugin_present?